requirements.txt
# An example requirements file. If your function has other dependencies,
# add them below
flask~=3.0.0
openai~=1.3.0
google-cloud-secret-manager
python-dotenv~=1.0.0